src/main/java/be/woutschoovaerts/mollie/serializer/MoneySerializer.java
package be.woutschoovaerts.mollie.serializer;

import com.fasterxml.jackson.core.JsonGenerator;
import com.fasterxml.jackson.databind.JsonSerializer;
import com.fasterxml.jackson.databind.SerializerProvider;

import java.io.IOException;
import java.math.BigDecimal;
import java.math.RoundingMode;

public class MoneySerializer extends JsonSerializer<BigDecimal> {
    @Override
    public void serialize(BigDecimal value, JsonGenerator jsonGenerator, SerializerProvider serializers) throws IOException {
        BigDecimal scaledValue = value.setScale(2, RoundingMode.HALF_UP);
        jsonGenerator.writeObject(scaledValue);
    }
}
